Move test_has_names to its own file
That way it can be run even on libraries that fail the hypothesis sanity
checks.
