+++ array_api_tests/test_signatures.py
@@ -32,9 +32,8 @@ from hypothesis.strategies import DataObject
 from . import dtype_helpers as dh
 from . import hypothesis_helpers as hh
 from . import xps
-from ._array_module import _UndefinedStub, mod as xp, mod_name
-from .stubs import (array_attributes, array_methods, category_to_funcs,
-                    extension_to_funcs, EXTENSIONS)
+from ._array_module import _UndefinedStub, mod as xp
+from .stubs import array_methods, category_to_funcs, extension_to_funcs
 from .typing import Array, DataType
 
 pytestmark = pytest.mark.ci
@@ -252,32 +251,3 @@ def test_array_method_signature(stub: FunctionType, data: DataObject):
     assert hasattr(x, stub.__name__), f"{stub.__name__} not found in array object {x!r}"
     method = getattr(x, stub.__name__)
     _test_func_signature(method, stub, array=x)
-
-has_name_params = []
-for ext, stubs in extension_to_funcs.items():
-    for stub in stubs:
-        has_name_params.append(pytest.param(ext, stub.__name__))
-for cat, stubs in category_to_funcs.items():
-    for stub in stubs:
-        has_name_params.append(pytest.param(cat, stub.__name__))
-for meth in array_methods:
-    has_name_params.append(pytest.param('array_method', meth.__name__))
-for attr in array_attributes:
-    has_name_params.append(pytest.param('array_attribute', attr))
-
-# This is a very basic test to see what names are defined in a library. It
-# does not even require functioning hypothesis array_api support.
-@pytest.mark.parametrize("category, name", has_name_params)
-def test_has_names(category, name):
-    if category in EXTENSIONS:
-        ext_mod = getattr(xp, category)
-        assert hasattr(ext_mod, name), f"{mod_name} is missing the {category} extension function {name}()"
-    elif category.startswith('array_'):
-        # TODO: This would fail if ones() is missing.
-        arr = xp.ones((1, 1))
-        if category == 'array_attribute':
-            assert hasattr(arr, name), f"The {mod_name} array object is missing the attribute {name}"
-        else:
-            assert hasattr(arr, name), f"The {mod_name} array object is missing the method {name}()"
-    else:
-        assert hasattr(xp, name), f"{mod_name} is missing the {category} function {name}()"
